Web: Editor: fixed duplication of global styles on each request
